Split training functions to ralf_train.py file. Updated requirements to include version numbers. SOURCES.txt to include ralf_train.py
