+++ ralf.egg-info/SOURCES.txt
@@ -2,6 +2,7 @@ README.md
 setup.py
 ralf/__init__.py
 ralf/ralf.py
+ralf/ralf_train.py
 ralf.egg-info/PKG-INFO
 ralf.egg-info/SOURCES.txt
 ralf.egg-info/dependency_links.txt
