+++ requirements.txt
@@ -1,10 +1,12 @@
-pandas
-openai
-peft
-transformers
-scikit-learn
-datasets
-evaluate
-google.generativeai
-humanize
-nltk
+datasets==4.0.0
+evaluate==0.4.5
+google-generativeai==0.8.5
+humanize==4.12.3
+nltk==3.9.1
+openai==1.100.0
+pandas==2.3.1
+peft==0.17.0
+psutil==7.0.0
+scikit-learn==1.7.1
+streamlit==1.48.1
+transformers==4.55.2
